split up state
