+++ src/frontend/components/check-balance/step-3/view-wallet-history.tsx
@@ -15,7 +15,9 @@ export class ViewWalletHistory extends React.Component<IViewWalletHistoryProps, 
     render(): JSX.Element {
         return (
             <div>
-                <WalletHistory/>
+                <WalletHistory
+                    state={this.props.state}
+                />
             </div>
         );
     }
