+++ src/frontend/components/contribute/step-1/contribute-start.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { contributeStartContent as content } from '../../../data/text-data';
 import { ProveYouAreHuman } from '../../../common/prove-you-are-human';
-import { setState, incrementStep } from '../../../state';
+import { setState, incrementStep } from '../../../state/index';
 import { State } from '../../../models';
 
 interface IContributeStartProps {
